requirements.txt
# PeerTube uploader dependencies
requests>=2.0.0
python-dotenv>=0.19.0